API improvements for dynamic usages (#117)
In light of recent experience report using this module with a dynamic
schema, this PR proposes several changes to make such work easier. It
also puts additional validation into initialization, so we can fail
faster and (hopefully) with better error messages.

1. Don't make user provide a type resolver. If user does not provide a
type resolver with a custom schema, use a smarter default that can use
dynamic message types and then fallback to the global registry for
anything unknown (including both extensions and message types).
2. Validate during initialization that the given type resolver can
actually resolve all of the methods' request and response types.

There is also a small change in here that is not strictly related to
dynamic usage but is instead related to proxying an RPC protocol to
REST: if a request comes in via RPC protocol for a method that has no
HTTP mapping and the target protocol is REST, this will now return a
"404 not found" status code. It previously could result in a
nil-dereference panic.