Support JSON name query parameters (#133)
Allows for query parameters to be specified by the fields JSON name.
This is supported by other implementations. Only enable JSON names for
query parameters and not within rule descriptors. By default encoding to
REST streams will now set query params as the JSON representation. This
is more consistent with the JSON body representation of the fields.

Fixes #131

---------

Signed-off-by: Edward McFarlane <emcfarlane@buf.build>