Update to latest golang.org/x/net (#136)
This indirectly updates to `golang.org/x/crypto` v0.27.0. This addresses
a potential vulnerability in versions of `golang.org/x/crypto` earlier
than v0.24.0 (previously, this repo indirectly depended on v0.21.0).

The vulnerability details are here:
https://security.snyk.io/vuln/SNYK-GOLANG-GOLANGORGXCRYPTOACMEAUTOCERT-7416897?_gl=1%2auism3r%2a_ga%2aMTg4NDYwMTUzMy4xNzE3NTE3MDA0%2a_ga_X9SH3KP7B4%2aMTcyNjg0MTA3OS42LjEuMTcyNjg0MTU2NC4wLjAuMA..

This repo was not actually susceptible to this vuln since it has no
codepath through which the vulnerable package was used. (In fact, the
vulnerable version of `golang.org/x/crypto` does not even appear in
`go.sum`, indicating that no package in this repo, even indirectly, uses
that module.) So it's actually a false positive. However, updating the
dependency is still good hygiene.